+++ examples/mjs/index.mjs
@@ -66,9 +66,11 @@ import fs from 'fs';
    * Post local files to the Turbo service.
    */
   console.log('Posting raw file to Turbo service...');
-  const filePath = new URL('../files/0_kb.txt', import.meta.url).pathname;
+  const filePath = new URL('../files/1KB_file', import.meta.url).pathname;
+  const fileSize = fs.statSync(filePath).size;
   const uploadResult = await turboAuthClient.uploadFile({
     fileStreamFactory: () => fs.createReadStream(filePath),
+    fileSizeFactory: () => fileSize,
     signal: AbortSignal.timeout(10_000), // cancel the upload after 10 seconds
   });
   console.log(JSON.stringify(uploadResult, null, 2));