+++ tests/turbo.node.test.ts
@@ -5,6 +5,7 @@ import { expect } from 'chai';
 import fs from 'fs';
 import { Readable } from 'node:stream';
 
+import { USD } from '../src/common/currency.js';
 import { JWKInterface } from '../src/common/jwk.js';
 import {
   TurboAuthenticatedClient,
@@ -12,6 +13,7 @@ import {
 } from '../src/common/turbo.js';
 import { TurboFactory } from '../src/node/factory.js';
 import { jwkToPublicArweaveAddress } from '../src/utils/base64.js';
+import { FailedRequestError } from '../src/utils/errors.js';
 
 describe('Node environment', () => {
   describe('TurboFactory', () => {
@@ -81,8 +83,7 @@ describe('Node environment', () => {
 
     it('getPriceForFiat()', async () => {
       const { winc } = await turbo.getWincForFiat({
-        amount: 1000, // 10 USD
-        currency: 'usd',
+        amount: USD(10), // $10.00 USD
       });
       expect(winc).to.not.be.undefined;
       expect(+winc).to.be.greaterThan(0);
@@ -97,6 +98,7 @@ describe('Node environment', () => {
 
         const response = await turbo.uploadSignedDataItem({
           dataItemStreamFactory: () => signedDataItem.getRaw(),
+          dataItemSizeFactory: () => signedDataItem.getRaw().length,
         });
         expect(response).to.not.be.undefined;
         expect(response).to.have.property('fastFinalityIndexes');
@@ -113,6 +115,7 @@ describe('Node environment', () => {
 
         const response = await turbo.uploadSignedDataItem({
           dataItemStreamFactory: () => Readable.from(signedDataItem.getRaw()),
+          dataItemSizeFactory: () => signedDataItem.getRaw().length,
         });
         expect(response).to.not.be.undefined;
         expect(response).to.have.property('fastFinalityIndexes');
@@ -129,10 +132,40 @@ describe('Node environment', () => {
         const error = await turbo
           .uploadSignedDataItem({
             dataItemStreamFactory: () => signedDataItem.getRaw(),
+            dataItemSizeFactory: () => signedDataItem.getRaw().length,
             signal: AbortSignal.timeout(0), // abort the request right away
           })
           .catch((err) => err);
         expect(error).to.be.instanceOf(CanceledError);
+      });
+
+      it('should return FailedRequestError for incorrectly signed data item', async () => {
+        const jwk = await Arweave.crypto.generateJWK();
+        const signer = new ArweaveSigner(jwk);
+        const signedDataItem = createData('signed data item', signer, {});
+        // not signed
+        const error = await turbo
+          .uploadSignedDataItem({
+            dataItemStreamFactory: () => signedDataItem.getRaw(),
+            dataItemSizeFactory: () => signedDataItem.getRaw().length,
+          })
+          .catch((err) => err);
+        expect(error).to.be.instanceOf(FailedRequestError);
+        expect(error.message).to.contain('Invalid Data Item');
+      });
+    });
+
+    describe('createCheckoutSession()', () => {
+      it('should properly get a checkout session', async () => {
+        const { adjustments, paymentAmount, quotedPaymentAmount, url } =
+          await turbo.createCheckoutSession({
+            amount: USD(10),
+            owner: '43-character-stub-arweave-address-000000000',
+          });
+        expect(adjustments).to.deep.equal([]);
+        expect(paymentAmount).to.equal(1000);
+        expect(quotedPaymentAmount).to.equal(1000);
+        expect(url).to.be.a('string');
       });
     });
   });
@@ -140,12 +173,14 @@ describe('Node environment', () => {
   describe('TurboAuthenticatedNodeClient', () => {
     let jwk: JWKInterface;
     let turbo: TurboAuthenticatedClient;
+    let address: string;
 
     before(async () => {
       jwk = await Arweave.crypto.generateJWK();
       turbo = TurboFactory.authenticated({
         privateKey: jwk,
       });
+      address = await Arweave.init({}).wallets.jwkToAddress(jwk);
     });
 
     it('getBalance()', async () => {
@@ -155,9 +190,11 @@ describe('Node environment', () => {
 
     describe('uploadFile()', () => {
       it('should properly upload a Readable to turbo', async () => {
-        const filePath = new URL('files/0_kb.txt', import.meta.url).pathname;
+        const filePath = new URL('files/1KB_file', import.meta.url).pathname;
+        const fileSize = fs.statSync(filePath).size;
         const response = await turbo.uploadFile({
           fileStreamFactory: () => fs.createReadStream(filePath),
+          fileSizeFactory: () => fileSize,
         });
         expect(response).to.not.be.undefined;
         expect(response).to.not.be.undefined;
@@ -168,14 +205,55 @@ describe('Node environment', () => {
       });
 
       it('should abort the upload when AbortController.signal is triggered', async () => {
-        const filePath = new URL('files/0_kb.txt', import.meta.url).pathname;
+        const filePath = new URL('files/1KB_file', import.meta.url).pathname;
+        const fileSize = fs.statSync(filePath).size;
         const error = await turbo
           .uploadFile({
             fileStreamFactory: () => fs.createReadStream(filePath),
+            fileSizeFactory: () => fileSize,
             signal: AbortSignal.timeout(0), // abort the request right away
           })
-          .catch((err) => err);
+          .catch((error) => error);
         expect(error).to.be.instanceOf(CanceledError);
+      });
+
+      it('should return a FailedRequestError when the file is larger than the free limit and wallet is underfunded', async () => {
+        const filePath = new URL('files/1MB_file', import.meta.url).pathname;
+        const fileSize = fs.statSync(filePath).size;
+        const error = await turbo
+          .uploadFile({
+            fileStreamFactory: () => fs.createReadStream(filePath),
+            fileSizeFactory: () => fileSize,
+          })
+          .catch((error) => error);
+        expect(error).to.be.instanceOf(FailedRequestError);
+        expect(error.message).to.contain('Insufficient balance');
+      });
+    });
+
+    it('getPriceForFiat() with a bad promo code', async () => {
+      const error = await turbo
+        .getWincForFiat({
+          amount: USD(10), // $10.00 USD
+          promoCodes: ['BAD_PROMO_CODE'],
+        })
+        .catch((error) => error);
+      expect(error).to.be.instanceOf(FailedRequestError);
+      // TODO: Could provide better error message to client. We have error messages on response.data
+      expect(error.message).to.equal('Failed request: 400: Bad Request');
+    });
+
+    describe('createCheckoutSession()', () => {
+      it('should fail to get a checkout session with a bad promo code', async () => {
+        const error = await turbo
+          .createCheckoutSession({
+            amount: USD(10), // 10 USD
+            owner: address,
+            promoCodes: ['BAD_PROMO_CODE'],
+          })
+          .catch((error) => error);
+        expect(error).to.be.instanceOf(FailedRequestError);
+        expect(error.message).to.equal('Failed request: 400: Bad Request');
       });
     });
   });